requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
numpy>=1.24.0
opencv-python-headless>=4.8.0   # headless = no GUI deps, saves ~150 MB
Pillow>=10.0.0

# ── TensorFlow Lite ────────────────────────────────────────────────────────────
# Install the official wheel for RPi 4 (armv7l / aarch64):
#   pip install tflite-runtime  (lightweight, no full TF)
#tflite-runtime>=2.14.0

# ── GPIO / Hardware ────────────────────────────────────────────────────────────
RPi.GPIO>=0.7.1
picamera2>=0.3.12               # modern Pi-camera stack

# ── Database ───────────────────────────────────────────────────────────────────
# SQLite is in stdlib; aiosqlite for async access
aiosqlite>=0.19.0

# ── HTTP / API ─────────────────────────────────────────────────────────────────
httpx>=0.27.0                   # async HTTP client (replaces requests)
tenacity>=8.2.0                 # retry logic

# ── Async ──────────────────────────────────────────────────────────────────────
# asyncio is stdlib; uvloop gives ~2x speed on Linux
uvloop>=0.19.0

# ── Optional TTS (text-to-speech) ──────────────────────────────────────────────
# pyttsx3>=2.90                 # offline TTS – uncomment if speaker available

# ── Logging / Config ───────────────────────────────────────────────────────────
python-dotenv>=1.0.0
